+++ mockall_derive/src/mock.rs
@@ -221,17 +221,19 @@ fn gen_mock_method(mod_ident: Option<&syn::Ident>,
     let abi = &sig.abi;
     let fn_token = &sig.decl.fn_token;
     let ident = &sig.ident;
-    let (ig, tg, wc) = sig.decl.generics.split_for_impl();
-    let merged_g = merge_generics(&generics, &sig.decl.generics);
     let meth_types = method_types(sig, Some(generics));
+    let merged_g = merge_generics(&generics, &meth_types.expectation_generics);
     let inputs = &meth_types.inputs;
     let output = &meth_types.output;
     let attrs = format_attrs(meth_attrs);
 
     // First the mock method
-    quote!(#attrs #meth_vis #constness #unsafety #asyncness #abi
-           #fn_token #ident #ig (#inputs) #output #wc)
-        .to_tokens(&mut mock_output);
+    {
+        let (ig, _, wc) = sig.decl.generics.split_for_impl();
+        quote!(#attrs #meth_vis #constness #unsafety #asyncness #abi
+               #fn_token #ident #ig (#inputs) #output #wc)
+            .to_tokens(&mut mock_output);
+    }
 
     let sub_name = if let Some(s) = sub {
         format!("{}_", s)
@@ -267,12 +269,15 @@ fn gen_mock_method(mod_ident: Option<&syn::Ident>,
         }
     }
 
-    let tg = if meth_types.is_static || !sig.decl.generics.params.is_empty() {
+    let (ig, ex_tg, wc) = meth_types.expectation_generics.split_for_impl();
+    let is_generic_method = !meth_types.expectation_generics.params.is_empty();
+    //let tg = ex_tg;
+    let tg = if meth_types.is_static || is_generic_method {
         // For generic and static methods only, the trait's generic parameters
         // become generic parameters of the method.
         merged_g.split_for_impl().1
     } else {
-        tg
+        ex_tg
     };
     let call_turbofish = tg.as_turbofish();
     if meth_types.is_static {
@@ -302,7 +307,7 @@ fn gen_mock_method(mod_ident: Option<&syn::Ident>,
             Span::call_site());
         let lt = syn::Lifetime::new("'guard", Span::call_site());
         let ltd = syn::LifetimeDef::new(lt);
-        let mut g = sig.decl.generics.clone();
+        let mut g = meth_types.expectation_generics.clone();
         g.params.push(syn::GenericParam::Lifetime(ltd.clone()));
         let merged_g = merge_generics(&generics, &g);
         let (ig, _, _) = g.split_for_impl();
@@ -368,7 +373,7 @@ fn gen_struct<T>(mock_ident: &syn::Ident,
         let attrs = format_attrs(&meth.borrow().attrs);
         let method_ident = &meth.borrow().sig.ident;
         let meth_types = method_types(&meth.borrow().sig, Some(generics));
-        let inputs = &meth_types.inputs;
+        //let inputs = &meth_types.inputs;
         let expect_obj = &meth_types.expect_obj;
         let expectations = &meth_types.expectations;
         let altargs = &meth_types.altargs;
@@ -377,9 +382,9 @@ fn gen_struct<T>(mock_ident: &syn::Ident,
         let output = &meth_types.output;
 
         let expect_vis = expectation_visibility(&meth.borrow().vis, 2);
-        let meth_generics = &meth.borrow().sig.decl.generics;
+        //let meth_generics = &meth.borrow().sig.decl.generics;
         let mut macro_g = TokenStream::new();
-        let merged_g = merge_generics(&generics, &meth_generics);
+        let merged_g = merge_generics(&generics, &meth_types.expectation_generics);
         if ! merged_g.params.is_empty() {
             merged_g.split_for_impl().1.to_tokens(&mut macro_g)
         } else {
@@ -388,7 +393,8 @@ fn gen_struct<T>(mock_ident: &syn::Ident,
         }
 
         let ecode = expectation(&attrs, &expect_vis, Some(&mock_ident),
-            &meth_ident, &merged_g, &inputs, output, &altargs, &matchexprs);
+            &meth_ident, &merged_g, &meth_types.expectation_inputs, output,
+            &altargs, &matchexprs);
         ecode.to_tokens(&mut mod_body);
 
         if meth_types.is_static {